Improvised the README with the current state of the project
After adding the issues to the repository, I changed the README to reflect what is the current state of the project and to delivered it to a state where it is more independent than it was before. That is, the improvement and development of the project should not be let alone with a team that changed/left very frenquently over the time and, in other perspective, the responsability of maintain it should not be left for only a person - an effort is to make this project well documented and open for anyone in the nucleus to grab where it was left on and improve it onwards, even by a little.