Fixed the application and should be ready for improvements
The app is now ready for improvements, being this commit the last changes in the code, hopefully.

It was changed the way that the app is called internally, changing the primeira_aplicacao to hs_app, fix some broken links used to load some of the images used in the main page's events and, lastly, the package name was updated in both platforms - Android and iOS.

As for the .gitignore file, was added some folders that weren't essential for the future of this project: web, windows, macos and linux.