AppBar improved!
Now the application topbar appears to have more functionalities and, with some changes in the issue #4, it doesn't look so thinner (at least for now).

In terms of fucntionalities, it launches the HS's website when the user presses on the HS's logo at the center; and it launches the main user page when the user presses on the icon with his/her image.

Since these were the changes requested for now, the PR is now complete and it should be closed. At least, once again, for now.