Added the CI to the project
The project now has CI, which build a release version of the app in either Android and iOS, when the master branch gets merged or gets a pull request.

Since this is the first time using CI in this project, there may be some new commits testing this new addition. If not, that means that everything worked just fine!